requirements.txt
cutadapt>=4.0
matplotlib>=3.7
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
scikit-learn>=1.3
scipy>=1.11
